Update train hash calculation to include passanger activity
Add pick up and drop off to the train hash calculation. This is due to
the rare case where a train overlay may have the _exact_ same timings as
the base schedule as well as the same day mask (binary mask of which
days of the week the train runs), but have different passanger activities.

The current behaviour will merge the two schedules together, meaning
some info is lost. The hash changes means they are kept are two distinct
schedules